Fix constprop inference for varargs OpaqueClosure (#39972)
OpaqueClosure are marked as vararg (or not) at construction time
rather than in the method. As a result, we need to chain this
information through to the cache. We may want to refactor this
code to deal with the vararg-ness (or not) of a particular
definition one level above this. We already keep the cache
in a form that represents the vararg tuple explicitly.
However, for now this fixes things for now (and also
prevents Cthulhu from getting confused upon encountering
these).