Fix opaque closure codegen ABI (#40001)
At the moment our ABI for specfun opaque closures is that the
opaque closure itself remains a GC tracked pointer out of which
the function itself implicitly loads the environment and the world
age. In the future we may want to improve this ABI, but it seems
fine for now. However, this ABI wasn't quite implemented correctly
after we've started turning on inference of OpaqueClosure where
the first argument of the MethodInstance's linfo actually stores
the type of the closure environment rather than the closure itself.
Fix this by manually forcing the signature determination code to
just treat that argument as `Any`, though of course in the future
we may want to do something fancier.