Prevent PCRE from needing a rebuild/reconf.
This commit reorders a couple touches in deps/pcre.mkl. Those touches are
intended to prevent patch application from triggering rebuild/reconf.
The present ordering doesn't quite succeed in that objective; patch
application triggers rebuild. On systems with compatible autotools,
the rebuild succeeds and the build is nonetheless happy. On systems
with incompatible versions of autotools, however, the build fails for
need of a reconf. The reordering in this commit prevents the rebuild.

Co-Authored-By: Rob Vermaas <rob.vermaas@gmail.com>