Implement OC captures getfield elim
This was originally in #40005, but we switched the representation
of accessing the OC captures from a sepcial intrinsic to inserting
an additional `getfieled(oc, :captures)`, which I thought obviated
the need for extra getfield elim support, but that's of course not
true, since getfield elim knows nothing about :new_opaque_closure.
Bring back that opimization and add a test, so we notice in the
future.