Restore moving GlobalRef out of argument position
This was dropped in #40066. I thought this was there because of
the side-effect that we used to allow in GlobalRefs, but we also
need to avoid inlining it into a PhiNode, and there are data
race concerns as well, so put back the outlining.
Fixes verifier complaints in `make debug` mode.