adding default argument value to reset!(stateful)
I think that most of the time when one calls `reset!(::Stateful)` it will be to simply restart from the begin without changing the iterated collection. So I figured that would avoid `reset!(sf, sf.itr)`. 
I also added the use of `reset!` to the documentation examples because it's not documented anywhere. This way users know of its existence.