RFC: frontend support for named OpaqueClosures
One gotcha is that defining a second opaque closure with the same name as the first will just define a new opaque closure. This is of course different to how method definitions usually work, but since opaque closures don't do dispatch this seems the most sensible to me. Another thing to think about is whether we want to make the binding a constant if the opaque closure is defined in global scope.