Replace _iszero by iszerodefined && iszero (#38789)
This PR defines an iszerodefined function that custom types implements in case iszero is well-defined even if zero is not defined.

Take 2 of #36194 after it was reverted in #38764. What was missing in #36194 is that in case iszero is not working for a given type, we should drop the structure and fall back to an Array.