Fix signed integer overflow in `mapreduce_impl` (#38643)
This PR solves #38627 as regards the code in reducedim.jl, reduce.jl and reinterpretarray.jl.

It does not yet solve all the relevant issues with missing.jl: in the code there for _mapreduce and mapreduce_impl, there are still several instances of the construct:
```
while i <= ilast
    ...
    i += 1
end
```
This overflows when ilast == typemax(Int). More seriously, inside the while loops there is often `@inbounds A[i]`, which is then an out-of-bounds read!