Add references to Performance Tips and Workflow Tips to Getting Started.
I suspect that some people go to julialang.org, click 'Learn', click
'getting started guide', start reading, and never get to Performance Tips
(or Workflow Tips) because they come quite late in the manual.

(An alternative would be to link to the introduction which already has a link to Performance Tips.
However, I like that the Getting Started page quickly gets to the REPL,
so I don't want to change that.)