[cli] Provide `.type` and `.size` annotations for symbols (#40575)
On ELF platforms, `ld` can print out warnings that certain symbols do
not have a type or size annotation.  This adds the annotations to the
`DEBUGINFO` and `EXPORT` helper macros, which are where this information
was put for the windows assembler.  With this patch, we should eliminate
`ld` warnings such as:

```
warning: type and size of dynamic symbol `jl_symbol' are not defined
```