AbstractInterpreter: refactor a bit `may_optimize` interface
Currently `OptimizationState` can be introduced even when `may_optimize`
returns `false`, and it can lead to a problem in `ir_to_codeinf!` call,
that takes `OptimizationState` and expects it to hold optimized `IRCode,
which doesn't necessarily happen when the optimization passes are turned
off by an external `AbstractInterpreter`.

With this PR, `OptimizationState` is introduced only when the optimization
is turned on, and the problem described above won't happen anymore.