IncrementalCompact: handle insertion by OldSSAValue before the insert pt (#41114)
This is a case that doesn't currently come up in our base transformation
passes. However, it's a well defined thing to want to do and I needed
it in Diffractor. That said, as more packages are starting to do compiler-y
things, we probably want to figure out a way to decouple general compiler
utilities from Core.Compiler, but in the meantime, just make this work.