fix equality of QRCompactWY
Equality for `QRCompactWY` did not ignore the subdiagonal entries of
`T` leading to nondeterministic behavior. Perhaps `T` should be
directly stored as `UpperTriangular` in `QRCompactWY`, but that seems
potentially breaking.

This is pulled out from #41228, since this change should be less
controversial than the other changes there and this particular bug just
came up in ChainRules again.