RFC: strip spaces after backslash-escaped newline
This was discussed on Slack with @StefanKarpinski and @BioTurboNick and
the general consensus was that stripping all spaces following a
backslash-escaped newline would make more sense than the current
behavior. One subtle distinction here is whether this stripping should
happen before or after triple-quoted strings. Right now it happens after
detenting so

```julia
"""
  a\
b"""
```

produces `"  ab"`, whereas if it would be the other way around, escaped
newlines would not affect detenting so the spaces preceding the `a`
would be removed.