Fix #41299 - Dropped error case in :> special case inliner
We have custom inference for :>, which prevents the :> method
itself from being inferred. As a result, it needs to be inlined
manually. However, this inlining was overeager and was dropping
an error check when the result was inferred to `Const`. Fix that
by verifying that the inlined <: call would have been nothrow.