Export jl_get_pgcstack from libjulia (#41320)
The pgcstack symbol is special because embedders are supposed to
define it themselves in localexec tls, and then access it that way.
However, there is another case of of libraries that link against
julia, but are not embedders, but rather plugins to be loaded later
(e.g. CxxWrap). These do need access to the pgcstack, but do not
have a priori access to the localexec of the embedder. Thus,
re-export the `jl_get_pgcstack` function that provides access
to our internal getter for the pgcstack.

(cherry picked from commit e469a1e4da7e7356257107d174b63fb5206bee04)