Make addprocs() safe for reuse
It seems common to say `addprocs()` to just get the number of processes to match the number of cores. And then you run the script again... and you get 2nprocs(). And 3nprocs()... I see this as a user trap with an easy fix. If `addprocs()` is "safe", then `addprocs()` at the top of a parallel script will always be smart and give you "the number of cores you probably want". This quick fix changes `addprocs()` to do just that.