deprecate unsafe_length for simply length (#40382)
This seems to be a fairly arbitrary case for throwing exceptions, when
the user might often use this value in arithmetic afterwards, which is
not checked. It leads to awkward complexity in the API however, where it
may be unclear which function to reach for, with no particular
justification for why a particular usage is "safe". And it inhibits
optimization and performance due to the additional checks and error
cases (and is not even entirely type-stable).

(cherry picked from commit 3eefaf0a52d1f537c512282a3027ead8e5e4f44e)