[build] Automatically set `USE_BINARYBUILDER_CSL=0` when local CSL is new (#41645)
We ship our own compiler support libraries to ensure a minimum level of
support for BB-built libraries, however certain distros provide very
bleeding-edge compilers.  This can be a problem if we download an
_older_ `libstdc++.so` and forcibly link against that when launching
Julia, as when Julia itself is built with the local `g++`, it may use
symbols that don't exist in the BB-sourced `libstdc++.so`.

To address this, we default to not using BB-sourced CSLs if the
`libstdc++.so` that the native compiler would use contains a symbol that
our BB-sourced CSLs do not have.  We use the monotonically-climbing
`GLIBCXX_3.4.XX` version symbols for this purpose; encoding the "next"
version number within `deps/csl.mk`, and triggering the BB-avoidance if
that version exists within the system-provided `libstdc++.so`.

(cherry picked from commit 27c0291f7fbecd811a3f9eb62b1159cff6b61e2a)