optimizer: Refactor out redundant `nargs` parameter (#41699)
These functions have taken `nargs` parameters for a long time
and through many refactors. However, at this point, having this
parameter makes little sense. In fact, we were passing it in three
different places here:
 1. As an explicit argument
 2. As a field of the OptimizationState
 3. Implicitly via the Method def of the MethodInstance in the
    OptimizationState.

Worse, we were inconsistent about whether the first argument counted
as an argument, so there was +/- 1 between these values. Clean that
all up by deleting cases 1 and 2 and retriving it directly from the
MethodInstance in the two places that need it.