Fix read! for types with size not a power of 2
Currently, the fast path in `read!(::IO, ::AbstractArray{T})` checks for
`isbitstype(T)` and whether the array is some nice type, then calls
`unsafe_read` with `sizeof` for the number of bytes. This works great in
the majority of cases, unless `sizeof(T)` is not a power of 2. The
aligned size is what's used in `sizeof`, which overestimates the number
of bytes required to read from the stream. To fix this, we can adjust
the condition of the fast path to also determine whether `sizeof`
matches the aligned size.