Get rid of unnecessary use of static parameter in ccall
`ccall` has some special magic to instantiate type parameters in argument
and return types. However, this is not particularly robust. It is the only place
that doesn't use `Expr(:static_parameter)` to access type parameters and
the applicability is limited. Moreover, it has semantic problems, because the
C signature doesn't actually change depending on the static parameter type
and the ccall should really represent the C signature.

Get rid of one unnecessary instance of this pattern I happened to stumble over
(though there's a couple more in the codebase).