handle the condition in `in` when `step` equals 0
Using `in` when `step(r::AbstractRange) == 0` results in an error.

Handle this case by using boolean logic for both the definitions to
which this case applies (`x::Real` and `x::AbstractChar`).
First, check if the step equals zero and if `x` (the value) equals the
reference value before performing the invalid `mod` operation.

Signed-off-by: Anant Thazhemadam <anant.thazhemadam@gmail.com>