Add special `@` character to `LOADER_BUILD_DEP_LIBS` and friends
This allows us to mark a library as `"special"` so that the loader
doesn't attempt to open it blindly, but rather interprets it as a list
of positional arguments, that it knows how to interpret.  We strictly
require exactly the number of special libraries, to help avoid errors in
the future as we add to this list.