Fix #41975 - Dropped typecheck in GotoIfNot
Recall the reproducer from the issue:
```
julia> f() = (if nothing; end; unsafe_load(Ptr{Int}(0)))
f (generic function with 1 method)

julia> f()
Unreachable reached at 0x7fb33bb50090

signal (4): Illegal instruction
in expression starting at REPL[13]:1
unsafe_load at ./pointer.jl:105 [inlined]
unsafe_load at ./pointer.jl:105 [inlined]
```

There were actually two places where we were dropping the
GotoIfNot, one in type annotation after inference, one in
SSA conversion. The one in SSA conversion was structural:
When both branches target the same jump destination, the
GotoIfNot would be dropped. This was fine in general, except
that as shown above, GotoIfNot can actually itself have
a side effect, namely throwing a type error if the condition
is not a boolean. Thus in order to actually drop the node
we need to prove that the error check does not fire.

The reason we want to drop the GotoIfNot node here is
that IRCode has an invariant that every basic block is
in the predecessor list only once (otherwise PhiNodes
would have to carry extra state regarding which branch
they refer to).

To fix this, on option would be to create a new expr
(e.g. `boolcheck`) that works like `:throw_undef_if_not`.
However, on second thought, I figured we might as well
just re-use GotoIfNot for this, since we'd otherwise
have to duplicate all the checks for side effects and
add codegen, etc., so this does the following:

  The `dest` label in GotoIfNot becomes optional. If set
  to 0, the only thing the GotoIfNot does is to perform
  the boolcheck. It is also no longer considered a terminator
  for SSA construction, either. The regular code for
  determining side-effects of GotoIfNot will determine
  the bool-ness of the condition and delete the GotoIfNot
  if applicable.

This also aligns with another PR I'm working on, which
will add an optional `else` label to GotoIfNot (in
the direction of #41476). Hopefully with this, the
two paths will then be symmetric.