Add a `@compile` directive to disable interpreter
We currently use a `while false; end` hack to disable to interpreter,
and this is getting more widely used outside of base. Unfortunately,
improvements to the interpreter that would allow us to interpret loops
would defeat the intention. This allows developers to annotate blocks
with `@compile` to specify that they want to force the compiler to run
on this block.