Clarify `mkpath` docstring
I had thought `mkpath("test/file.txt")` would make a directory `test` and that's it (https://github.com/rofinn/FilePathsBase.jl/issues/129#issuecomment-915197193). But it also creates `file.txt` as a directory. @omus pointed out this could be clarified in the `mkpath` docstring so this is my attempt.

I removed the first bit of the example with `mkdir` because I felt that it was confusing. `mkdir` and `mkpath` are similar and if you are on the `mkpath` docstring and it starts with `mkdir` it could be easy to get mixed up. The point of that code was to start in an empty directory I believe, so I just did `cd(mktempdir())` instead.