Use `unwrapva` in MethodError candidates
I found an example of a method declaration with an argument
typed as `a::Vararg`, and this triggered an error in showing method
candidates. It's always bad when there's an error triggered when handling
an error. This PR makes the code more robust against such problems.

It's worth noting that this type of method declaration
shouldn't be used:  it is better as `a...`. The only time you should
explicitly use `Vararg` is when you are specifying both `T` and
`N` in `Vararg{T,N}`.