Move `SHA` to be an external stdlib (#41370)
For the past few years, we have had `SHA` as a vendored copy, however
improvements have been made both to the vendored copy here in base as
well as to the registered copy living within `JuliaCrypto`.  This
eliminates the vendored copy and sets `SHA` up as an external stdlib
just like `Pkg`, `Tar`, and all of their friends, as well as merges the
improvements made to both into the external repository.