specialize `hastypemax` on `Bool`
Since `Bool` is not included in `Base.BitIntegerType`, `hastypemax(Bool)` falls back to the slow generic version. This reduces the performance of `digits!` with Boolean target vector (for `base=2`) significantly. Specializing `hastypemax` on `Bool` fixes this performance bug. 
In one of my applications, where `digits!` should only take a small fraction of the total runtime, fixing this improves the performance by more than 10%.