change locks to avoid capturing pthread_self value
It was observed in TSAN that we might get incorrectly locked mutexes
when this code is inlined, due to pthread_self being normally marked
const. Therefore, switch most nogc-variant locks to use standard system
mutexes instead. We may want to work towards switching over other lock
types too, but note that they will need to integrate with gc-safepoints,
which would be a design effort for later.