inference: make `widenconst` always return `Type`
The primary purpose of this commit is to improve the handling of
`TypeVar`s and `Vararg`s.

This commit tries to modify all places where we call `widenconst` on
those non-`Type` objects and make inference work if `widenconst` doesn't
accept them. Now `widenconst` is ensured to return `Type` always (and as
a result we can safely call e.g. `widenconst(x) <: widenconst(y)`).