eliminate unbound `TypeVar`s on `argtypes` construction
This commit eliminates unbound `TypeVar`s from `argtypes` in order to
make the analysis much easier down the road.

At runtime, only `Type{...}::DataType` can contain invalid type parameters,
but we may have arbitrary malformed user-constructed type arguments given
at inference entries.
So we will replace only the malformed `Type{...}::DataType` with `Type`
and simply replace other possibilities with `Any`.
The replacement might not be that accurate, but an unbound `TypeVar` is
really invalid in anyway.

Like the change added on `arrayref_tfunc`, now we may be able to remove
some `isa(x, TypeVar)`/`has_free_typevars` predicates used in various
places, but it is left as an exercise for the reader.