Eliminate usage of `rm` command-line program
Note that we use `recursive=true` as we might encounter a `libjulia-codegen.X.Y.dylib.dSYM` directory.  Although deleting that directory is not critical for this test, not throwing an exception is, so we just remove it as well.  Also, this test would silently fail to test what we want it to if run out of a build directory, so we first teach it to dynamically find the location of `libjulia-codegen`, then secondly, cause it to fail if it does not remove any `libjulia-codegen` libraries.