Cleanup `RPATH` settings
Previously, we needed to provide both `usr/lib` and `usr/lib/julia` as
elements on the `RPATH`, because `libjulia` lived in `usr/lib`, and
needed to be able to find dependent libraries both within a build tree
(when most libraries live in `usr/lib`) and within an install tree (when
most libraries live in `usr/lib/julia`).

Nowadays, everything is either in `usr/lib` or in `usr/lib/julia` with
the exception of `libjulia`, but since it's really `libjulia-internal`
that does all the `dlopen()`'ing, we can simply use `$$ORIGIN` (or the
equivalent `@loader_path/` on macOS) and completely ignore the rest of
the RPATH shenanigans we do.