Partially revert #42919
The change to RPATH setting for macOS seems to have improved the
situation for users on Monterey, but the other changes in #42919 that
affect other platforms seem to have caused regressions on musl Linux
(#42940) and FreeBSD (#42944). The easiest thing to do here is to keep
the changes from #42919 as they apply to macOS but revert the others.