codegen: add missing initialization for PhiC nodes (#43029)
Our Phi handling assumes that it can references undefined memory, and
get back legal results, but our PhiC nodes were not initialized, so the
Phi node might see uninitialized results, and then cause the GC to
crash. This was observed in PkgEval on the PoreMatMod.jl package to
occur in recent Julia versions and master.