loading: define module before include, to solve a minor data-race
Passing the UUID via global state can be a data-race, and passing it via
TLS seems more awkward than necessary. Instead, we define that `require`
implicitly makes the module it expects. We make the module a baremodule
initially, until the user calls `module Name` (or `using Base` and
defines `include` and `eval`). The module call also (re)sets the scope
of the `__init__` call, to preserve the prior behavior and order of it.