make our locks fast (#42964)
Now that we have the ability to express atomics (particularly the write barrier of the Task object), it is time for us to speed up and simplify our locks.

This PR does need someone to verify the synchronizations are correct, but I think this version is actually quite simple. There is a single byte that represents the state of the lock:

0x0 : unlocked
0x1 : locked (without other waiters)
0x2 : locked with waiters
And then when we have the locked bit set, we can manipulate any of the other data (e.g. owner and reentrancy) without needing any atomic work, and if we don't hold that bit, we should not write to any other field.