Friendlier more literal names for token kinds
Parser source which tests token kinds is easier to read when those
tokens are represented as string literals; the flisp code uses
character literals to great effect for this purpose. In Julia code we
can do even better by introducing a string macro for these purposes.
Introduce TK"]" for the token kind of the ']' token and a TokenKinds
module to hold the mapping between Symbols and kinds.