Relocate `llvm-version.h` in codegen to avoid a warning (#43181)
Our `llvm-version.h` defines `LLVM_ENABLE_STATS` if it's not already
defined. However, we're `#include`ing that file in `codegen.cpp` before
we `#include` the LLVM headers that define it. That means it gets
defined twice, triggering a `-Wmacro-redefined` warning. If we move the
inclusion of our file to below the inclusion of LLVM's headers, we just
use their definition and avoid a warning.