Stop hardcoding SOVERSION for stdlibs
Hardcoding versions is actually not needed as the build already creates
symlinks from libXXX.so to libXXX.so.Y.Z. Since Julia does not always
need a particular version, it is better to use the link which is created
based on the one which was available at build time.
It also reduced the number of places to edit when upgrading to a new SOVERSION.
This was already the case for several libraries on Windows.