use has_fma intrinsic for math (#43259)
This removes all bad `fma` checks in `Base` and replaces them with fma
multiversioned ones. It also adds `Base.Math.two_mul` which gives the
fastest compensated multiplication available regardless of whether the
user has `fma` hardware. This is what should be used when possible for
compensated multiplication in the future.