Restore valgrind suppressions file
contrib/valgrind-julia.supp was removed in #38970, but it is still referenced
[in the Valgrind devdocs](https://docs.julialang.org/en/v1/devdocs/valgrind/).
Either the file should be restored, or the reference should be removed
from the documentation.

Here, I have chosen the option of restoring the file, as I think it
provides some value.  I have also updated it such that it does indeed
suppress the warning (#4533) on the latest master.